+++ lib/src/analysis/divisional_chart_service.dart
@@ -31,7 +31,8 @@ class DivisionalChartService {
         '_${rashiChart.ascendant.hashCode}'
         '_${planetsHash}_${type.name}';
 
-    return _cache.putIfAbsent(key, () => _computeDivisionalChart(rashiChart, type));
+    return _cache.putIfAbsent(
+        key, () => _computeDivisionalChart(rashiChart, type));
   }
 
   VedicChart _computeDivisionalChart(